Backend/deepseek-api-ms/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
requests>=2.30.0
tenacity>=8.2.2
